+++ hyperon_das/das.py
@@ -1,23 +1,11 @@
-import json
-from abc import ABC, abstractmethod
-from typing import Any, Dict, List, Optional, Set, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 
-import requests
-from hyperon_das_atomdb import WILDCARD, AtomDB
+from hyperon_das_atomdb import AtomDB
 from hyperon_das_atomdb.adapters import InMemoryDB, RedisMongoDB
-from hyperon_das_atomdb.exceptions import AtomDoesNotExist, NodeDoesNotExist
 
-from hyperon_das.cache import AndEvaluator, LazyQueryEvaluator, ListIterator, QueryAnswerIterator
-from hyperon_das.client import FunctionsClient
-from hyperon_das.decorators import retry
-from hyperon_das.exceptions import (
-    InvalidDASParameters,
-    InvalidQueryEngine,
-    QueryParametersException,
-    UnexpectedQueryFormat,
-)
+from hyperon_das.engines import LocalQueryEngine, RemoteQueryEngine
+from hyperon_das.exceptions import InvalidDASParameters, InvalidQueryEngine
 from hyperon_das.logger import logger
-from hyperon_das.utils import Assignment, QueryAnswer
 
 
 class DistributedAtomSpace:
@@ -50,10 +38,6 @@ class DistributedAtomSpace:
                 details=f'query_engine={query_engine_parameter}',
             )
 
-    def _error(self, exception: Exception):
-        logger().error(str(exception))
-        raise exception
-
     def get_atom(self, handle: str) -> Union[Dict[str, Any], None]:
         """
         Retrieve information about an Atom using its handle.
@@ -68,15 +52,16 @@ class DistributedAtomSpace:
             Union[Dict, None]: A dictionary containing detailed Atom information
 
         Raises:
-            ValueError: If an invalid output format is provided.
+            AtomDoesNotExist: If invalid parameter is provided.
 
         Example:
             >>> result = das.get_atom(handle="af12f10f9ae2002a1607ba0b47ba8407")
             >>> print(result)
             {
-                "handle": "af12f10f9ae2002a1607ba0b47ba8407",
-                "type": "Concept",
-                "name": "human"
+                'handle': 'af12f10f9ae2002a1607ba0b47ba8407',
+                'composite_type_hash': 'd99a604c79ce3c2e76a2f43488d5d4c3',
+                'name': 'human',
+                'named_type': 'Concept'
             }
         """
         return self.query_engine.get_atom(handle)
@@ -91,13 +76,10 @@ class DistributedAtomSpace:
             node_name (str): The name of the specific node being queried.
 
         Returns:
-            Union[Dict, None]: A dictionary containing atom information of the node
+            Union[Dict, None]: A dictionary containing detailed node information
 
         Raises:
-            ValueError: If an invalid output format is provided.
-
-        Note:
-            If the specified node does not exist, a warning is logged and None is returned.
+            NodeDoesNotExist: If invalid parameters are provided.
 
         Example:
             >>> result = das.get_node(
@@ -106,9 +88,10 @@ class DistributedAtomSpace:
                 )
             >>> print(result)
             {
-                "handle": "af12f10f9ae2002a1607ba0b47ba8407",
-                "type": "Concept",
-                "name": "human"
+                'handle': 'af12f10f9ae2002a1607ba0b47ba8407',
+                'composite_type_hash': 'd99a604c79ce3c2e76a2f43488d5d4c3',
+                'name': 'human',
+                'named_type': 'Concept'
             }
         """
         return self.query_engine.get_node(node_type, node_name)
@@ -120,24 +103,41 @@ class DistributedAtomSpace:
 
         Args:
             link_type (str): The type of the link being queried.
-            targets (List[str], optional): A list of target identifiers that the link is associated with.
-                Defaults to None.
+            link_targets (List[str]): A list of target identifiers that the link is associated with.
 
         Returns:
-            Union[Dict, None]: A dictionary containing atom information of the link
+            Union[Dict, None]: A dictionary containing detailed link information
 
         Raises:
-            ValueError: If an invalid output format is provided.
+            LinkDoesNotExist: If invalid parameters are provided.
 
         Note:
             If the specified link or targets do not exist, the method returns None.
 
         Example:
+            >>> human_handle = das.get_node_handle('Concept', 'human')
+            >>> monkey_handle = das.get_node_handle('Concept', 'monkey')
             >>> result = das.get_link(
                     link_type='Similarity',
-                    targets=['human', 'monkey'],
+                    link_targets=[human_handle, monkey_handle],
                 )
             >>> print(result)
+            {
+                'handle': 'bad7472f41a0e7d601ca294eb4607c3a',
+                'composite_type_hash': 'ed73ea081d170e1d89fc950820ce1cee',
+                'is_toplevel': True,
+                'composite_type': [
+                    'a9dea78180588431ec64d6bc4872fdbc',
+                    'd99a604c79ce3c2e76a2f43488d5d4c3',
+                    'd99a604c79ce3c2e76a2f43488d5d4c3'
+                ],
+                'named_type': 'Similarity',
+                'named_type_hash': 'a9dea78180588431ec64d6bc4872fdbc',
+                'targets': [
+                    'af12f10f9ae2002a1607ba0b47ba8407',
+                    '1cdffc6b0b89ff41d68bec237481d1e1'
+                ]
+            }
 
         """
         return self.query_engine.get_link(link_type, link_targets)
@@ -156,17 +156,14 @@ class DistributedAtomSpace:
         Args:
             link_type (str): The type of links being queried.
             target_types (List[str], optional): The type(s) of targets being queried. Defaults to None.
-            targets (List[str], optional): A list of target identifiers that the links are associated with.
+            link_targets (List[str], optional): A list of target identifiers that the links are associated with.
                 Defaults to None.
 
         Returns:
             Union[List[str], List[Dict]]: A list of dictionaries containing detailed information of the links
 
-        Raises:
-            ValueError: If an invalid output format is provided or if the provided parameters are invalid.
-
         Example:
-            >>> result = obj.get_links(
+            >>> result = das.get_links(
                     link_type='Similarity',
                     target_types=['Concept', 'Concept']
                 )
@@ -242,10 +239,7 @@ class DistributedAtomSpace:
                     }
                 ]
             })
-            >>> query_params = {
-                "toplevel_only": False,
-                "return_type": QueryOutputFormat.ATOM_INFO,
-            }
+            >>> query_params = {"toplevel_only": False}
             >>> q1 = {
                 "atom_type": "link",
                 "type": "Expression",
@@ -263,7 +257,25 @@ class DistributedAtomSpace:
             }
             >>> result = das.query(q1, query_params)
             >>> print(result)
-            [{'handle': 'dbcf1c7b610a5adea335bf08f6509978', 'type': 'Expression', 'template': ['Expression', 'Symbol', ['Expression', 'Symbol', 'Symbol']], 'targets': [{'handle': '963d66edfb77236054125e3eb866c8b5', 'type': 'Symbol', 'name': 'Test'}, {'handle': '233d9a6da7d49d4164d863569e9ab7b6', 'type': 'Expression', 'template': ['Expression', 'Symbol', 'Symbol'], 'targets': [{'handle': '963d66edfb77236054125e3eb866c8b5', 'type': 'Symbol', 'name': 'Test'}, {'handle': '9f27a331633c8bc3c49435ffabb9110e', 'type': 'Symbol', 'name': '2'}]}]}]
+            [
+                {
+                    'handle': 'dbcf1c7b610a5adea335bf08f6509978',
+                    'type': 'Expression',
+                    'template': ['Expression', 'Symbol', ['Expression', 'Symbol', 'Symbol']],
+                    'targets': [
+                        {'handle': '963d66edfb77236054125e3eb866c8b5', 'type': 'Symbol', 'name': 'Test'},
+                        {
+                            'handle': '233d9a6da7d49d4164d863569e9ab7b6',
+                            'type': 'Expression',
+                            'template': ['Expression', 'Symbol', 'Symbol'],
+                            'targets': [
+                                {'handle': '963d66edfb77236054125e3eb866c8b5', 'type': 'Symbol', 'name': 'Test'},
+                                {'handle': '9f27a331633c8bc3c49435ffabb9110e', 'type': 'Symbol', 'name': '2'}
+                            ]
+                        }
+                    ]
+                }
+            ]
         """
         return self.query_engine.query(query, parameters)
 
@@ -334,7 +346,7 @@ class DistributedAtomSpace:
                     'type': 'Reactome',
                     'name': 'Reactome:R-HSA-164843',
                 }
-            >>> db.add_node(node_params)
+            >>> das.add_node(node_params)
         """
         return self.backend.add_node(node_params)
 
@@ -349,8 +361,6 @@ class DistributedAtomSpace:
             link_params (Dict[str, Any]): A dictionary containing link parameters. It should have the following keys:
                 - 'type': The type of the link.
                 - 'targets': A list of target elements.
-            toplevel: boolean flag to indicate toplevel links
-                i.e. links which are not nested inside other links.
 
         Returns:
             Dict[str, Any]: The information about the added link, including its unique key and other details.
@@ -372,7 +382,7 @@ class DistributedAtomSpace:
                         },
                     ],
                 }
-            >>> db.add_link(link_params)
+            >>> das.add_link(link_params)
         """
         return self.backend.add_link(link_params)
 
@@ -380,242 +390,3 @@ class DistributedAtomSpace:
         """Clear all data"""
         self.backend.clear_database()
         logger().debug('The database has been cleaned.')
-
-
-class QueryEngine(ABC):
-    @abstractmethod
-    def get_atom(self, handle: str) -> Union[Dict[str, Any], None]:
-        ...
-
-    @abstractmethod
-    def get_node(self, node_type: str, node_name: str) -> Union[Dict[str, Any], None]:
-        ...
-
-    @abstractmethod
-    def get_link(self, link_type: str, targets: List[str]) -> Union[Dict[str, Any], None]:
-        ...
-
-    @abstractmethod
-    def get_links(
-        self, link_type: str, target_types: List[str] = None, link_targets: List[str] = None
-    ) -> Union[List[str], List[Dict]]:
-        ...
-
-    @abstractmethod
-    def query(
-        self,
-        query: Dict[str, Any],
-        parameters: Optional[Dict[str, Any]] = {},
-    ) -> List[Dict[str, Any]]:
-        ...
-
-    @abstractmethod
-    def count_atoms(self) -> Tuple[int, int]:
-        ...
-
-
-class LocalQueryEngine(QueryEngine):
-    def __init__(self, backend, kwargs: Optional[dict] = None) -> None:
-        self.local_backend = backend
-
-    def _recursive_query(
-        self,
-        query: Union[Dict[str, Any], List[Dict[str, Any]]],
-        mappings: Set[Assignment] = None,
-        parameters: Optional[Dict[str, Any]] = None,
-    ) -> QueryAnswerIterator:
-        if isinstance(query, list):
-            sub_expression_results = [
-                self._recursive_query(expression, mappings, parameters) for expression in query
-            ]
-            return AndEvaluator(sub_expression_results)
-        elif query["atom_type"] == "node":
-            try:
-                atom_handle = self.local_backend.get_node_handle(query["type"], query["name"])
-                return ListIterator(
-                    [QueryAnswer(self.local_backend.get_atom_as_dict(atom_handle), None)]
-                )
-            except NodeDoesNotExist:
-                return ListIterator([])
-        elif query["atom_type"] == "link":
-            matched_targets = []
-            for target in query["targets"]:
-                if target["atom_type"] == "node" or target["atom_type"] == "link":
-                    matched = self._recursive_query(target, mappings, parameters)
-                    if matched:
-                        matched_targets.append(matched)
-                elif target["atom_type"] == "variable":
-                    matched_targets.append(ListIterator([QueryAnswer(target, None)]))
-                else:
-                    self._error(
-                        UnexpectedQueryFormat(
-                            message="Query processing reached an unexpected state",
-                            details=f'link: {str(query)} link target: {str(query)}',
-                        )
-                    )
-            return LazyQueryEvaluator(query["type"], matched_targets, self, parameters)
-        else:
-            self._error(
-                UnexpectedQueryFormat(
-                    message="Query processing reached an unexpected state",
-                    details=f'query: {str(query)}',
-                )
-            )
-
-    def _to_link_dict_list(self, db_answer: Union[List[str], List[Dict]]) -> List[Dict]:
-        if not db_answer:
-            return []
-        flat_handle = isinstance(db_answer[0], str)
-        answer = []
-        for atom in db_answer:
-            if flat_handle:
-                handle = atom
-                arity = -1
-            else:
-                handle, targets = atom
-                arity = len(targets)
-            answer.append(self.local_backend.get_atom_as_dict(handle, arity))
-        return answer
-
-    def get_atom(self, handle: str) -> Union[Dict[str, Any], None]:
-        try:
-            return self.local_backend.get_atom(handle)
-        except AtomDoesNotExist:
-            return None
-
-    def get_node(self, node_type: str, node_name: str) -> Union[Dict[str, Any], None]:
-        try:
-            node_handle = self.local_backend.node_handle(node_type, node_name)
-            return self.local_backend.get_atom(node_handle)
-        except AtomDoesNotExist:
-            return None
-
-    def get_link(self, link_type: str, link_targets: List[str]) -> Union[Dict[str, Any], None]:
-        try:
-            link_handle = self.local_backend.link_handle(link_type, link_targets)
-            return self.local_backend.get_atom(link_handle)
-        except AtomDoesNotExist:
-            return None
-
-    def get_links(
-        self, link_type: str, target_types: List[str] = None, link_targets: List[str] = None
-    ) -> Union[List[str], List[Dict]]:
-        if target_types is not None and link_type != WILDCARD:
-            db_answer = self.local_backend.get_matched_type_template([link_type, *target_types])
-        elif link_targets is not None:
-            db_answer = self.local_backend.get_matched_links(link_type, link_targets)
-        elif link_type != WILDCARD:
-            db_answer = self.local_backend.get_matched_type(link_type)
-        else:
-            self._error(ValueError("Invalid parameters"))
-
-        return self._to_link_dict_list(db_answer)
-
-    def query(
-        self,
-        query: Dict[str, Any],
-        parameters: Optional[Dict[str, Any]] = {},
-    ) -> List[Dict[str, Any]]:
-        logger().debug(
-            {
-                'message': '[DistributedAtomSpace][query] - Start',
-                'data': {'query': query, 'parameters': parameters},
-            }
-        )
-        query_results = self._recursive_query(query, parameters)
-        logger().debug(f"query: {query} result: {str(query_results)}")
-        answer = []
-        for result in query_results:
-            answer.append(result.subgraph)
-        return answer
-
-    def count_atoms(self) -> Tuple[int, int]:
-        return self.local_backend.count_atoms()
-
-    def commit(self):
-        self.local_backend.commit()
-
-
-class RemoteQueryEngine(QueryEngine):
-    def __init__(self, backend, kwargs):
-        self.local_query_engine = LocalQueryEngine(backend, kwargs)
-        host = kwargs.get('host')
-        port = kwargs.get('port')
-        if not host:
-            raise InvalidDASParameters(message='Send `host` parameter to connect in a remote DAS')
-        url = self._connect_server(host, port)
-        self.remote_das = FunctionsClient(url)
-
-    @retry(attempts=5, timeout_seconds=120)
-    def _connect_server(self, host: str, port: Optional[str] = None):
-        port = port or '8081'
-        openfaas_uri = f'http://{host}:{port}/function/query-engine'
-        aws_lambda_uri = f'http://{host}/prod/query-engine'
-        url = None
-        if self._is_server_connect(openfaas_uri):
-            url = openfaas_uri
-        elif self._is_server_connect(aws_lambda_uri):
-            url = aws_lambda_uri
-        return url
-
-    def _is_server_connect(self, url: str) -> bool:
-        try:
-            response = requests.request(
-                'POST',
-                url=url,
-                data=json.dumps({"action": "ping", "input": {}}),
-                timeout=10,
-            )
-        except Exception as e:
-            return False
-        if response.status_code == 200:
-            return True
-        return False
-
-    def get_atom(self, handle: str) -> Dict[str, Any]:
-        local = self.local_query_engine.get_atom(handle)
-        if not local:
-            return self.remote_das.get_atom(handle)
-
-    def get_node(self, node_type: str, node_name: str) -> Dict[str, Any]:
-        local = self.local_query_engine.get_node(node_type, node_name)
-        if not local:
-            return self.remote_das.get_node(node_type, node_name)
-
-    def get_link(self, link_type: str, link_targets: List[str]) -> Dict[str, Any]:
-        local = self.local_query_engine.get_link(link_type, link_targets)
-        if not local:
-            return self.remote_das.get_link(link_type, link_targets)
-
-    def get_links(
-        self, link_type: str, target_types: List[str] = None, link_targets: List[str] = None
-    ) -> Union[List[str], List[Dict]]:
-        local = self.local_query_engine.get_links(link_type, target_types, link_targets)
-        if not local:
-            return self.remote_das.get_links(link_type, target_types, link_targets)
-
-    def query(
-        self,
-        query: Dict[str, Any],
-        parameters: Optional[Dict[str, Any]] = {},
-    ) -> List[Dict[str, Any]]:
-        query_scope = parameters.get('query_scope', 'remote_only')
-        if query_scope == 'remote_only':
-            answer = self.remote_das.query(query, parameters)
-        elif query_scope == 'local_only':
-            answer = self.local_query_engine.query(query, parameters)
-        elif query_scope == 'local_and_remote':
-            # This type is not available yet
-            raise QueryParametersException
-        elif query_scope == 'synchronous_update':
-            self.commit()
-            answer = self.remote_das.query(query, parameters)
-        return answer
-
-    def count_atoms(self) -> Tuple[int, int]:
-        local_answer = self.local_query_engine.count_atoms()
-        remote_answer = self.remote_das.count_atoms()
-        return tuple([x + y for x, y in zip(local_answer, remote_answer)])
-
-    def commit(self):
-        return self.remote_das.commit_changes()
